refactor: Move configuration classes to common module
