+++ jkube-kit/common/src/main/java/org/eclipse/jkube/kit/common/JKubeConfiguration.java
@@ -11,7 +11,7 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.jkube.kit.config.image.build;
+package org.eclipse.jkube.kit.common;
 
 import java.io.File;
 import java.io.Serializable;
@@ -24,8 +24,6 @@ import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.eclipse.jkube.kit.config.image.RegistryConfig;
-import org.eclipse.jkube.kit.common.JavaProject;
 
 /**
  * @author roland
