+++ jkube-kit/config/service/src/main/java/org/eclipse/jkube/kit/config/service/BuildService.java
@@ -14,7 +14,7 @@
 package org.eclipse.jkube.kit.config.service;
 
 import org.eclipse.jkube.kit.config.image.ImageConfiguration;
-import org.eclipse.jkube.kit.config.image.RegistryConfig;
+import org.eclipse.jkube.kit.common.RegistryConfig;
 
 import java.util.Collection;
